Material - BasicMaterial과 Material로 분할. BasicMaterial이 color buffer를 가지고 있는 형태. 나머지는 Material이 들고 있다. #37
